Refactor current ModelStub util to its own file, add option to ModuleBuilder to set isDone attribute
